+++ csvkit/utilities/csvlook.py
@@ -24,6 +24,7 @@ class CSVLook(CSVKitUtility):
     def main(self):
         table = agate.Table.from_csv(
             self.input_file,
+            skip_lines=self.args.skip_lines,
             sniff_limit=self.args.sniff_limit,
             header=not self.args.no_header_row,
             column_types=self.get_column_types(),