+++ csvkit/utilities/csvsql.py
@@ -112,6 +112,7 @@ class CSVSQL(CSVKitUtility):
             try:
                 table = agate.Table.from_csv(
                     f,
+                    skip_lines=self.args.skip_lines,
                     sniff_limit=self.args.sniff_limit,
                     header=not self.args.no_header_row,
                     column_types=self.get_column_types(),