+++ csvkit/utilities/csvstat.py
@@ -132,6 +132,7 @@ class CSVStat(CSVKitUtility):
 
         table = agate.Table.from_csv(
             self.input_file,
+            skip_lines=self.args.skip_lines,
             sniff_limit=self.args.sniff_limit,
             header=not self.args.no_header_row,
             **self.reader_kwargs