+++ CHANGELOG.rst
@@ -4,6 +4,7 @@
 Improvements:
 
 * Add a :code:`--version` (:code:`-V`) flag.
+* Add a :code:`--skip-lines` option to skip initial lines (e.g. comments, copyright notices, empty rows).
 * :code:`-I` is the short option for :code:`--no-inference`.
 * :doc:`/scripts/csvjoin` supports :code:`--snifflimit` and :code:`--no-inference`.
 * :doc:`/scripts/csvstat` adds a :code:`--freq-count` option to set the maximum number of frequent values to display.